+++ src/AbstractContext.spec.ts
@@ -66,6 +66,7 @@ describe( module( "Carbon/AbstractContext" ), ():void => {
 		it( hasProperty(
 			INSTANCE,
 			"parentContext",
+			"Carbon.Context",
 			"The parent context provided in the constructor. " +
 			"If no context has provided, the property will be the singleton `Carbon.SDKContext.instance` of the class `Carbon.SDKContext.Class`."
 		), ():void => {